backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic>=2.8.0
python-multipart==0.0.12
qdrant-client>=1.10.0
openai>=1.50.0
python-dotenv==1.0.1